Use toggleable multi-select cheat list
