+++ swing/src/main/java/eu/rekawek/coffeegb/swing/SwingMenu.kt
@@ -53,6 +53,7 @@ import java.awt.event.MouseEvent
 import java.awt.event.WindowAdapter
 import java.awt.event.WindowEvent
 import java.io.File
+import javax.swing.DefaultListCellRenderer
 import javax.swing.DefaultListModel
 import javax.swing.JCheckBoxMenuItem
 import javax.swing.JFileChooser
@@ -359,13 +360,33 @@ class SwingMenu(
       }
 
       val cheatChoices =
-          CheckBoxList(
-              supportedCheats,
-              ::cheatLabel,
-              { cheat -> "${cheat.description()} (${cheat.code()})" },
-          )
+          ToggleSelectionList(supportedCheats)
       cheatChoices.visibleRowCount = minOf(12, supportedCheats.size)
       cheatChoices.selectedIndex = 0
+      cheatChoices.cellRenderer =
+          object : DefaultListCellRenderer() {
+            override fun getListCellRendererComponent(
+                list: JList<*>?,
+                value: Any?,
+                index: Int,
+                isSelected: Boolean,
+                cellHasFocus: Boolean,
+            ): Component {
+              val component =
+                  super.getListCellRendererComponent(
+                      list,
+                      value,
+                      index,
+                      isSelected,
+                      cellHasFocus,
+                  )
+              if (component is JLabel && value is CheatDatabase.Cheat) {
+                component.text = cheatLabel(value)
+                component.toolTipText = "${value.description()} (${value.code()})"
+              }
+              return component
+            }
+          }
       val scrollPane = JScrollPane(cheatChoices)
       scrollPane.preferredSize =
           Dimension(
@@ -471,10 +492,6 @@ class SwingMenu(
   }
 
   private fun moveListSelection(list: JList<*>, offset: Int) {
-    if (list is CheckBoxList<*>) {
-      list.moveActiveIndex(offset)
-      return
-    }
     if (list.model.size == 0) {
       return
     }
